- move createMealHelper to utils package. - move MockMeals to utils package.
Took 4 minutes
