+++ src/test/kotlin/data/CsvFileReaderTest.kt
@@ -8,7 +8,6 @@ import org.junit.jupiter.api.io.TempDir
 import java.io.File
 import java.io.FileNotFoundException
 
-
 class CsvFileReaderTest {
 
     @TempDir
