+++ src/test/kotlin/domain/usecase/GetEasyMealSuggestionUseCaseTest.kt
@@ -8,6 +8,7 @@ import org.example.domain.MealException
 import org.example.domain.repository.MealRepository
 import org.example.domain.usecase.GetEasyMealSuggestionUseCase
 import org.junit.jupiter.api.BeforeEach
+import utils.MockMeals
 import kotlin.test.Test
 
 class GetEasyMealSuggestionUseCaseTest {
