+++ src/test/kotlin/domain/usecase/GetMealWithHighCaloriesUseCaseTest.kt
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import utils.createMeal
+import utils.createNutrition
 
 class GetMealWithHighCaloriesUseCaseTest {
     private val mealRepository = mockk<MealRepository>(relaxed = true)
