+++ src/test/kotlin/domain/usecase/GetMealsByCountryUseCaseTest.kt
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import utils.createMeal
 import kotlin.test.assertEquals
 
 class GetMealsByCountryUseCaseTest {
