+++ src/test/kotlin/domain/usecase/GetSeafoodMealsUseCaseTest.kt
@@ -8,6 +8,7 @@ import org.example.domain.repository.MealRepository
 import org.example.domain.usecase.GetSeafoodMealsUseCase
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import utils.MockMeals
 
 class GetSeafoodMealsUseCaseTest {
