+++ src/test/kotlin/domain/usecase/SearchByMealNameUseCaseTest.kt
@@ -9,6 +9,7 @@ import org.example.domain.MealException
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import utils.createMeal
 import kotlin.collections.listOf
 import kotlin.test.assertEquals
